requirements.txt
google-api-python-client
google-auth-oauthlib
tinydb
